% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-main.R
\name{autoplot.wp_model}
\alias{autoplot.wp_model}
\title{Autoplot for win probability model object}
\usage{
\method{autoplot}{wp_model}(object, type = "grid", ...)
}
\arguments{
\item{object}{win probability model}

\item{type}{one of \code{grid} (default), \code{coefs}, and \code{round}}
}
\value{
A ggplot2 plot
}
\description{
Autoplot for win probability model object
}
