% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snd.R
\name{check_snd}
\alias{check_snd}
\title{Check the structure of SND}
\usage{
check_snd(snd)
}
\arguments{
\item{snd}{an SND}
}
\value{
a standard abort message (if applicable)
}
\description{
A SND has a standard, fixed structure, which includes the following.
\itemize{
\item dataset 1
\item dataset 2
\item dataset 3
\item ...
}
}
\details{
which a dataset includes
\itemize{
\item snd_factor (named factor)
\item snd_item (named item)
\item data (named data)
}

This function checks the structure of the SND and display a warning/ abort if...
\itemize{
\item There is no datasets (Part 1, warning only)
\item There is no factor (Part 2)
\item There is no item (Part 3)
\item There is no data (Part 4)
}
}
\keyword{internal}
