% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFile_xlsx.R
\name{checkFile_xlsx}
\alias{checkFile_xlsx}
\title{System tools: Check if a \code{.xlsx} file exists}
\usage{
checkFile_xlsx(xlsxFile)
}
\arguments{
\item{xlsxFile}{Directory of the xlsxFile}
}
\value{
a standard message (if applicable)
}
\description{
\code{snd} can be stored within the computer as a \code{.xlsx} file. This function checks if the file directory provided is \code{character(0)}, not existing, or has incorrect extensions (i.e. not \code{.xlsx}).
}
\keyword{internal}
