% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRW_matrix.R
\name{checkRW_matrix}
\alias{checkRW_matrix}
\title{System tools: Precheck for r/w data matrix}
\usage{
checkRW_matrix(mtx, mtxName)
}
\arguments{
\item{mtx}{The matrix}

\item{mtxName}{The name of the matrix}
}
\value{
A standard abort message, if applicable
}
\description{
Prior to read-in a matrix, the matrix must loosely resemble the final snd product. These functions check if the matrix fits those criteria to be transformed into a snd matrix.
}
\keyword{Internal}
