% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\alias{classify_factor}
\alias{classify_item}
\alias{classify_data}
\alias{classify_set}
\alias{classify_key}
\title{An easier way to add additional classes to an object}
\usage{
classify(x, class)

classify_factor(x)

classify_item(x)

classify_data(x)

classify_set(x)

classify_key(x)
}
\arguments{
\item{x}{Any object}

\item{class}{The class to add. This will always be in front of the original classes of \code{x}}
}
\value{
\code{x}, now also classed in \code{class}
}
\description{
Self-explanatory... just like the title
For \code{classify_key}, the key doesn't need to remove \code{@} in front. It will automatically remove it for you. However, if you did remove it, it still works
}
\examples{
classify(X, "snd_data")
}
\keyword{internal}
