% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_extract.R
\name{extract}
\alias{extract}
\title{Extracting a dataset in SND}
\usage{
extract(snd, dataset = NULL)
}
\arguments{
\item{snd}{the SND object.}

\item{dataset}{the names of datasets to be extracted, in character.}
}
\value{
An SND
}
\description{
For an SND to maintain it's structure, the data dataframe can not be read alone! This function extracts the dataset while still maintaining the SND structure (i.e. also extracting the factor and item matrixes).
People do not need to call the names of the factor matrix or item matrix while using this function, just type the name(s) of the dataset (e.g. \code{c("reference_1", "treatment_1")}) in the argument \code{dataset}.
By default (\code{NULL}), all datasets will be extracted (i.e. returning the original SND). If the name(s) in \code{dataset} is not found in the SND, the name will be ignored, and the extraction will continue.
}
\examples{
extract(snd, c("month1", "month3"))
}
