% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp-fit.R
\name{fit_wp_model}
\alias{fit_wp_model}
\alias{fit_naive_wp_model}
\title{Fit a win probability model}
\usage{
fit_wp_model(data, ...)

fit_naive_wp_model(data)
}
\arguments{
\item{data}{data.frame with at least the following columns: \code{side}, \code{is_post_plant}, \code{n_team_pre_activity}}
}
\value{
tlist of two data.frames under the names \code{pre} and \code{post}
}
\description{
Use situational variables to make a win probability "model" that is really just
a lookup table of emppirical win proportions given the situation (time independent).
}
