% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_forge_xlsx.R
\name{forge_xlsx}
\alias{forge_xlsx}
\title{Forge in an SND with xlsx}
\usage{
forge_xlsx(xlsxFile, sheet)
}
\arguments{
\item{xlsxFile}{Directory of the .xlsx file}

\item{sheet}{Name of sheets to be read it, in \code{character}. Note that only sheets containing data needs to be specified. By default (\code{NULL}), all available data will be read.}
}
\value{
A SND object
}
\description{
\strong{WARNING: In general, DO NOT use this function unless you have good reasons.}

This function forges an SND, with only the data matrix, following the rules set by the Notion document and in \code{snd::read_xlsx}. As this function fills in all the necessary components in a SND, this is a good first step to start formatting your xlsx sheets into SND-compatible, along with \code{snd::write_xlsx}. The data matrix that needs to be SND compatible will need to use the prefix "#data_". If a sheet name in \code{sheet} is unavailable, the function will abort.
}
\examples{
forge_xlsx(xlsxFile, sheet = NULL)
}
