% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sndkey_format_detect.R
\name{format_detect}
\alias{format_detect}
\title{Detect what format a vector should be.
The detection is based on how many NAs there are. The more NAs, the more unlikely the format supports it.}
\usage{
format_detect(column)
}
\arguments{
\item{column}{A column of items in data.frame}
}
\value{
a character of the expected formats
}
\description{
Detect what format a vector should be.
The detection is based on how many NAs there are. The more NAs, the more unlikely the format supports it.
}
\examples{
format_detect(column)
}
\keyword{internal}
