% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_key.R
\name{format_key}
\alias{format_key}
\alias{format_key_type}
\alias{format_key_item}
\alias{format_key_datatype}
\alias{format_key_factor}
\alias{format_key_label}
\title{Trial}
\usage{
format_key(key, mtx, mtxName)

format_key_type(mtx, mtxName)

format_key_item(mtx, mtxName)

format_key_datatype(mtx, mtxName)

format_key_factor(mtx, mtxName)

format_key_label(mtx, mtxName)
}
\arguments{
\item{mtxName}{}
}
\description{
Trial
}
\keyword{internal}
