% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_key2mtx.R, R/format_withkey.R,
%   R/format_withkey_datatype_use.R
\name{format_key2mtx}
\alias{format_key2mtx}
\alias{format_key2mtx.sndkey_factor}
\alias{format_key2mtx.sndkey_item}
\alias{format_key2mtx.sndkey_type}
\alias{format_key2mtx.sndkey_label}
\alias{format_key2mtx.sndkey_datatype}
\alias{format_key2mtx_datatype_use}
\alias{format_key2mtx_datatype_use.DTnumeric}
\alias{format_key2mtx_datatype_use.DTlogical}
\alias{format_key2mtx_datatype_use.DTinteger}
\alias{format_key2mtx_datatype_use.DTcharacter}
\alias{format_key2mtx_datatype_use.DTPOSIXct}
\alias{format_key2mtx_datatype_use.DTfactor}
\alias{format_key2mtx_datatype_use.DTflexible}
\title{System tools: Checking and Formatting other matrixs using keys during read-in.}
\usage{
format_key2mtx(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_factor}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_item}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_type}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_label}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_datatype}(key, formater, formatee, formateeName)

format_key2mtx(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_factor}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_item}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_type}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_label}(key, formater, formatee, formateeName)

\method{format_key2mtx}{sndkey_datatype}(key, formater, formatee, formateeName)

format_key2mtx_datatype_use(colItem, colName, mtxName)

\method{format_key2mtx_datatype_use}{DTnumeric}(colItem, colName, mtxName)

\method{format_key2mtx_datatype_use}{DTlogical}(colItem, colName, mtxName)

\method{format_key2mtx_datatype_use}{DTinteger}(colItem, colName, mtxName)

\method{format_key2mtx_datatype_use}{DTcharacter}(colItem, colName, mtxName)

\method{format_key2mtx_datatype_use}{DTPOSIXct}(colItem, colName, mtxName)

\method{format_key2mtx_datatype_use}{DTfactor}(colItem, colName, mtxName)

\method{format_key2mtx_datatype_use}{DTflexible}(colItem, colName, mtxName)
}
\arguments{
\item{key}{The key in \code{character}. Only 1 key will be accepted.}

\item{formater}{The matrix in \code{data.frame} to format the formatee. The matrix must contain the key.}

\item{formatee}{The matrix in \code{data.frame} to be formatted using formater's keys.}

\item{formateeName}{The name of the formatee in \code{character}.}
}
\value{
The formatee matrix, checked (and formatted) by formater matrix.

The formatee matrix, checked (and formatted) by formater matrix.
}
\description{
In SND, a \code{key} is a column within the matrix that can be used to check and modify itself, it's matrix and other matrices. This function uses a key within \code{formater}, to checks and modifies another matrix \code{formatee}, using a key. To modify it's own matrix, check the function \code{snd:::format_key}. To enhance security, only 1 key will be accepted.

In SND, a \code{key} is a column within the matrix that can be used to check and modify itself, it's matrix and other matrices. This function uses a key within \code{formater}, to checks and modifies another matrix \code{formatee}, using a key. To modify it's own matrix, check the function \code{snd:::format_key}, which you are encouraged to use prior to formatting other matrices. To enhance security, only 1 key will be accepted.
}
\keyword{internal}
