% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_matrix.R, R/format_tools.R
\name{format_matrix}
\alias{format_matrix}
\alias{format_matrix.snd_data}
\alias{format_matrix.snd_item}
\alias{format_matrix.snd_factor}
\alias{format_MakeKeyType}
\alias{format_shuffle}
\title{System tools: Formatting while doing read-in operations}
\usage{
format_matrix(mtx, mtxName)

\method{format_matrix}{snd_data}(mtx, mtxName)

\method{format_matrix}{snd_item}(mtx, mtxName)

\method{format_matrix}{snd_factor}(mtx, mtxName)

format_MakeKeyType(mtx, mtxName)

format_shuffle(mtx, mtxName)
}
\arguments{
\item{mtx}{The matrix}

\item{mtxName}{The name of the matrix}
}
\value{
A dataframe
}
\description{
These functions process the matrix and form them into the final product prior to formating them into a dataset.
}
\keyword{internal}
