% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_shuffle.R
\name{format_shuffle}
\alias{format_shuffle}
\title{System tools: Shuffle matrix columns.}
\usage{
format_shuffle(mtx, mtxName)
}
\arguments{
\item{mtx}{A matrix of any kind}

\item{mtxName}{Name of the matrix}
}
\description{
In \code{snd}, how the matrixes are read are importants as it dictates which keys work first. This program first shuffles all the matrix and determines what is read first by using the simple to use \code{snd:::key_priority} function. A higher priority, the faster it will be read.
}
\keyword{internal}
