% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sndkey_format_rename.R
\name{format_sndkeyFormat_rename}
\alias{format_sndkeyFormat_rename}
\title{System tools: Format items within the key \verb{@format}}
\usage{
format_sndkeyFormat_rename(col)
}
\arguments{
\item{col}{A vector of characters.}
}
\value{
The analysis dataframe, with \code{full} as the full name of the format
}
\description{
To ease data entry within \code{.xlsx} file, people are allowed to type in abbreviated formats into the \code{.xlsx} file. However, this is not beneficial for computers! This function reformats everything back to it's full name form.
}
\examples{
snd:::format_sndkeyFormat_rename(col = c("#c", "#i", "i", "c", "lol"))
#The last example `lol` should return `match` as `0L` and `fullname` as NA
}
\keyword{interal}
