% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_withkey.R
\name{format_withkey}
\alias{format_withkey}
\alias{format_withkey_factor}
\alias{format_withkey_item}
\alias{format_withkey_type}
\alias{format_withkey_label}
\alias{format_withkey_datatype}
\alias{format_withkey_datatype_use}
\alias{format_withkey_datatype_use.DTnumeric}
\alias{format_withkey_datatype_use.DTlogical}
\alias{format_withkey_datatype_use.DTinteger}
\alias{format_withkey_datatype_use.DTcharacter}
\alias{format_withkey_datatype_use.DTPOSIXct}
\alias{format_withkey_datatype_use.DTfactor}
\alias{format_withkey_datatype_use.DTflexible}
\title{Testing}
\usage{
format_withkey(key, formater, formatee, formateeName)

format_withkey_factor(formater, formatee, formateeName)

format_withkey_item(formater, formatee, formateeName)

format_withkey_type(formater, formatee, formateeName)

format_withkey_label(formater, formatee, formateeName)

format_withkey_datatype(formater, formatee, formateeName)

format_withkey_datatype_use(colItem, colName, mtxName)

\method{format_withkey_datatype_use}{DTnumeric}(colItem, colName, mtxName)

\method{format_withkey_datatype_use}{DTlogical}(colItem, colName, mtxName)

\method{format_withkey_datatype_use}{DTinteger}(colItem, colName, mtxName)

\method{format_withkey_datatype_use}{DTcharacter}(colItem, colName, mtxName)

\method{format_withkey_datatype_use}{DTPOSIXct}(colItem, colName, mtxName)

\method{format_withkey_datatype_use}{DTfactor}(colItem, colName, mtxName)

\method{format_withkey_datatype_use}{DTflexible}(colItem, colName, mtxName)
}
\arguments{
\item{formatee}{}
}
\description{
Testing
}
\keyword{internal}
