% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_matrix_by.R
\name{formatData_byItem}
\alias{formatData_byItem}
\alias{formatData_byFactor}
\title{System tools: Format data matrix by some other matrix}
\usage{
formatData_byItem(mtxData, mtxItem, mtxName)

formatData_byFactor(mtxData, mtxFactor, mtxName)
}
\arguments{
\item{mtxData}{The data matrix}

\item{mtxItem}{The item matrix}

\item{mtxName}{The name of the data matrix}

\item{mtxFactor}{The factor matrix}
}
\value{
A formatted Data matrix
}
\description{
A work around based on \code{snd:::format_matrix} and \code{snd:::format_key2mtx}
}
\keyword{internal}
