% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatRI_key2mtx.R,
%   R/formatRI_key2mtx_format_use.R, R/formatRI_key2mtx_sndkey_label.R
\name{formatRI_key2mtx}
\alias{formatRI_key2mtx}
\alias{formatRI_key2mtx.sndkey_factor}
\alias{formatRI_key2mtx.sndkey_item}
\alias{formatRI_key2mtx.sndkey_type}
\alias{formatRI_key2mtx.sndkey_format}
\alias{formatRI_key2mtx_format_use}
\alias{formatRI_key2mtx_format_use.DTcharacter}
\alias{formatRI_key2mtx_format_use.DTnumeric}
\alias{formatRI_key2mtx_format_use.DTinteger}
\alias{formatRI_key2mtx_format_use.DTlogical}
\alias{formatRI_key2mtx_format_use.DTPOSIXct}
\alias{formatRI_key2mtx.sndkey_label}
\title{System tools: Checking and Formatting other matrixs using keys during read-in.}
\usage{
formatRI_key2mtx(key, formater, formatee, formateeName)

\method{formatRI_key2mtx}{sndkey_factor}(key, formater, formatee, formateeName)

\method{formatRI_key2mtx}{sndkey_item}(key, formater, formatee, formateeName)

\method{formatRI_key2mtx}{sndkey_type}(key, formater, formatee, formateeName)

\method{formatRI_key2mtx}{sndkey_format}(key, formater, formatee, formateeName)

formatRI_key2mtx_format_use(format, colItem, colName, mtxName)

\method{formatRI_key2mtx_format_use}{DTcharacter}(format, colItem, colName, mtxName)

\method{formatRI_key2mtx_format_use}{DTnumeric}(format, colItem, colName, mtxName)

\method{formatRI_key2mtx_format_use}{DTinteger}(format, colItem, colName, mtxName)

\method{formatRI_key2mtx_format_use}{DTlogical}(format, colItem, colName, mtxName)

\method{formatRI_key2mtx_format_use}{DTPOSIXct}(format, colItem, colName, mtxName)

\method{formatRI_key2mtx}{sndkey_label}(key, formater, formatee, formateeName)
}
\arguments{
\item{key}{The key in \code{character}. Only 1 key will be accepted.}

\item{formater}{The matrix in \code{data.frame} to format the formatee. The matrix must contain the key.}

\item{formatee}{The matrix in \code{data.frame} to be formatted using formater's keys.}

\item{formateeName}{The name of the formatee in \code{character}.}
}
\value{
The formatee matrix, checked (and formatted) by formater matrix.
}
\description{
In SND, a \code{key} is a column within the matrix that can be used to check and modify itself, it's matrix and other matrices. This function uses a key within \code{formater}, to checks and modifies another matrix \code{formatee}, using a key. To modify it's own matrix, check the function \code{snd:::formatRI_key}. To enhance security, only 1 key will be accepted.
}
\keyword{internal}
