% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatRI_matrix.R
\name{formatRI_matrix}
\alias{formatRI_matrix}
\title{System tools: Formatting while doing read-in operations}
\usage{
formatRI_matrix(mtx, mtxName)
}
\arguments{
\item{mtx}{The matrix}

\item{mtxName}{The name of the matrix}
}
\value{
A dataframe
}
\description{
These functions process the matrix and form them into the final product prior to formating them into a dataset.
}
\keyword{internal}
