% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_getUniqueName.R
\name{getUniqueNames}
\alias{getUniqueNames}
\title{Obtain unique names from a list}
\usage{
getUniqueNames(name)
}
\arguments{
\item{name}{A character vector. Can be names in a list, or column names ...}
}
\value{
A character vector, with all unique names
}
\description{
SND does not tolerate identical names, especially for its datasets as xlsx files can not be saved with identical sheet names!
Feed in a list of names in the \code{name}, and it will pop out a list of names that are completely unique using suffixes.
}
\keyword{internal}
