% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab-key.R
\name{grab_keyHead}
\alias{grab_keyHead}
\alias{grab_keyTail}
\title{System tools: Grab properties of key}
\usage{
grab_keyHead(mtx, key)

grab_keyTail(mtx, key)
}
\arguments{
\item{mtx}{mtx itself}

\item{key}{the name of the key. Only accepts characters and only accept 1}
}
\value{
A string of characters
}
\description{
Properties of a key is specified and separated using "_". These functions find the "head" and "tail" part of the key.

The functions are as follow:
grab_keyHead: Obtain the head of the key
grab_keyTail: Obtain the tail of the key
}
