% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab-mtx.R
\name{grab_mtxKey}
\alias{grab_mtxKey}
\alias{grab_mtxFactor}
\alias{grab_mtxItem}
\title{System tools: Grab various sheet names from .xlsx file}
\usage{
grab_mtxKey(dataframe)

grab_mtxFactor(dataframe)

grab_mtxItem(dataframe)
}
\arguments{
\item{dataframe}{Dataframe itself}
}
\value{
A string of characters
}
\description{
SND requires a organized 2D data structure to represent multivariant data. Additionally, it requires some way that users can easily access and modify the data, even without using R. These functions below with a prefix \code{sys_grab_df} provides the first steps to analysis the organized data structure in a matrix, which is stored as dataframe for ease of use.

The functions are as follow:
grab_mtxKey: Obtain Keys that will form the matrix
grab_mtxFactor: Obtain Factors that will form the matrix
grab_mtxItem: Obtain Items that will form the matrix
}
