% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_grab-snd.R
\name{grab_sndset}
\alias{grab_sndset}
\alias{grab_sndfactor}
\alias{grab_snditem}
\alias{grab_snddata}
\alias{grab_sndos}
\title{System tools: Grab various snd objects from a list}
\usage{
grab_sndset(snd)

grab_sndfactor(snd)

grab_snditem(snd)

grab_snddata(snd)

grab_sndos(snd)
}
\arguments{
\item{snd}{}
}
\value{
A list of objects
}
\description{
System tools: Grab various snd objects from a list
}
