% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab-xlsx.R
\name{grab_xlsxFactor}
\alias{grab_xlsxFactor}
\alias{grab_xlsxItem}
\alias{grab_xlsxData}
\title{System tools: Grab various sheet names from .xlsx file}
\usage{
grab_xlsxFactor(xlsxFile)

grab_xlsxItem(xlsxFile)

grab_xlsxData(xlsxFile)
}
\arguments{
\item{xlsxFile}{Directory of the excel file}
}
\value{
A string of characters
}
\description{
SND requires a organized 2D data structure to represent multivariant data. Additionally, it requires some way that users can easily access and modify the data, even without using R. These functions below with a prefix \code{sys_grab_xls} provides the first steps to analysis the organized data structure in the \code{.xlsx} file.

The functions are as follow:
\itemize{
\item grab_xlsxFactor: Obtain sheets that will form \code{factor} matrix
\item grab_xlsxItem: Obtain sheets that will form \code{item} matrix
\item grab_xlsxData: Obtain sheets that will form \code{data} matrix
}
}
