% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_is.R
\name{is_snd}
\alias{is_snd}
\title{Test if an object is an object from \code{snd}}
\usage{
is_snd(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value. \code{TRUE} for SED data.
}
\description{
Returns \code{TRUE} for said item being a snd_xxx, and \code{FALSE} for all other objects, including lists.
}
