% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ISO8601.R
\name{read_ISO8601}
\alias{read_ISO8601}
\alias{write_ISO8601}
\title{Title}
\usage{
read_ISO8601(time, tzone = "")

write_ISO8601(time)
}
\arguments{
\item{time}{}

\item{tzone}{}
}
\description{
Title

Title
}
