% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAs.R
\name{nameAs}
\alias{nameAs}
\title{An easier way to add names into an object}
\usage{
nameAs(x, name)
}
\arguments{
\item{x}{Any object}

\item{name}{names of the item inside of \code{x}, in character.}
}
\value{
\code{x}, but with names
}
\description{
Self-explanatory... just like the title
}
\keyword{internal}
