% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_OSgrab.R
\name{OSgrab}
\alias{OSgrab}
\title{Grabs OS content of a particular key}
\usage{
OSgrab(snd, OSkey)
}
\arguments{
\item{snd}{snd}

\item{OSkey}{a character}
}
\value{
value of that key
}
\description{
OSkey refers to the names of items stored in the \code{snd_os}. This function grabs the content of that key.
}
\keyword{internal}
