% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_OSmodify.R
\name{OSmodify}
\alias{OSmodify}
\title{Modifies OS content}
\usage{
OSmodify(snd, OSkey, value)
}
\arguments{
\item{snd}{snd}

\item{OSkey}{a character}

\item{value}{the value of the \code{OSkey}}
}
\value{
\code{snd}, with OS modified
}
\description{
OSkey refers to the names of items stored in the \code{snd_os}. Modifing this \code{snd_os} will lead to behavior change in \code{snd} and \code{seal}.
}
\keyword{internal}
