% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snd-package.R
\docType{package}
\name{snd-package}
\alias{snd}
\alias{snd-package}
\title{snd: R package for modeling Call of Duty SnD win probability}
\description{
A framework for modeling win probability in Call of Duty (CoD) Search and Destroy (SnD). Enables visualization of continuous win probability within a round.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tonyelhabr/snd}
}

}
\author{
\strong{Maintainer}: Tony ElHabr \email{anthonyelhabr@gmail.com}

}
\keyword{internal}
