% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_split.R
\name{split}
\alias{split}
\title{Split dataset(s) in SND}
\usage{
split(snd, dataset = NULL)
}
\arguments{
\item{snd}{the SND object.}

\item{dataset}{the names of datasets to be extracted, in character.}
}
\value{
An SND
}
\description{
For an SND to maintain it's structure, the data dataframe can not be read alone! This function splits the dataset(s) while still maintaining the SND structure (i.e. also extracting the factor and item matrixes).
People do not need to call the names of the factor matrix or item matrix while using this function, just type the name(s) of the datasets (e.g. \code{c("reference_1", "treatment_1")}) in the argument \code{dataset}.
By default (\code{NULL}), all datasets will be extracted (i.e. returning the original SND). If the name(s) in \code{dataset} is not found in the SND, the name will be ignored, and the extraction will continue.
}
\examples{
split(snd, c("month1", "month3"))
}
