% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-abort_file.R
\name{sys_abort_fileWrongType}
\alias{sys_abort_fileWrongType}
\title{System tools: Standard abort message for files}
\usage{
sys_abort_fileWrongType(x, expected_format)
}
\arguments{
\item{x}{The argument holding the file directory.}
}
\value{
A standard abort message
}
\description{
System tools: Standard abort message for files
}
\keyword{internal}
