% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_abort_mtx.R, R/sys_abort_mtxMissingKey.R
\name{sys_abort_mtx}
\alias{sys_abort_mtx}
\alias{sys_abort_mtxMissingSelectedKey}
\title{System tools: Standard abort message for matrices}
\usage{
sys_abort_mtx(name)

sys_abort_mtxMissingSelectedKey(x, keys_missing, name)
}
\arguments{
\item{name}{Optional: name of the matrix}

\item{x}{The argument holding the matrix}
}
\value{
A standard abort message
}
\description{
System tools: Standard abort message for matrices
}
\keyword{internal}
