% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-abort_sht.R
\name{sys_abort_shtUnavailable}
\alias{sys_abort_shtUnavailable}
\title{System tools: Standard abort message for sheets}
\usage{
sys_abort_shtUnavailable(x, unavailable_sheets)
}
\arguments{
\item{x}{The argument holding the xlsx.}
}
\value{
A standard abort message
}
\description{
System tools: Standard abort message for sheets
}
\keyword{internal}
