% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-datatype_support.R
\name{sys_datatype_support}
\alias{sys_datatype_support}
\title{System tools: List out all supported datatypes in SND}
\usage{
sys_datatype_support(with_abbr = F)
}
\value{
A list of supported datatypes/ classes, in \code{character}
}
\description{
System tools: List out all supported datatypes in SND
}
\examples{
sys_datatype_support()
}
\keyword{internal}
