% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab-df.R
\name{sys_grab_dfKey}
\alias{sys_grab_dfKey}
\alias{sys_grab_dfFactor}
\alias{sys_grab_dfItem}
\title{System tools: Grab various sheet names from .xlsx file}
\usage{
sys_grab_dfKey(dataframe)

sys_grab_dfFactor(dataframe)

sys_grab_dfItem(dataframe)
}
\arguments{
\item{dataframe}{Dataframe itself}
}
\value{
A string of characters
}
\description{
SND requires a organized 2D data structure to represent multivariant data. Additionally, it requires some way that users can easily access and modify the data, even without using R. These functions below with a prefix \code{sys_grab_df} provides the first steps to analysis the organized data structure in a matrix, which is stored as dataframe for ease of use.

The functions are as follow:
sys_grab_dfKey: Obtain Keys that will form the matrix
sys_grab_dfFactor: Obtain Factors that will form the matrix
sys_grab_dfItem: Obtain Items that will form the matrix
}
\keyword{internal}
