% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab-key.R
\name{sys_grab_keyHead}
\alias{sys_grab_keyHead}
\alias{sys_grab_keyTail}
\title{System tools: Grab properties of key}
\usage{
sys_grab_keyHead(mtx, key)

sys_grab_keyTail(mtx, key)
}
\arguments{
\item{mtx}{mtx itself}

\item{key}{the name of the key. Only accepts characters and only accept 1}
}
\value{
A string of characters
}
\description{
Properties of a key is specified and separated using "_". These functions find the "head" and "tail" part of the key.

The functions are as follow:
sys_grab_KeyHead: Obtain the head of the key
sys_grab_keyTail: Obtain the tail of the key
}
\keyword{Internal}
\keyword{internal}
