% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab-xls.R
\name{sys_grab_xlsfactor}
\alias{sys_grab_xlsfactor}
\alias{sys_grab_xlsitem}
\alias{sys_grab_xlsdata}
\title{System tools: Grab various sheet names from .xlsx file}
\usage{
sys_grab_xlsfactor(xlsxFile)

sys_grab_xlsitem(xlsxFile)

sys_grab_xlsdata(xlsxFile)
}
\arguments{
\item{xlsxFile}{Directory of the excel file}
}
\value{
A string of characters
}
\description{
SND requires a organized 2D data structure to represent multivariant data. Additionally, it requires some way that users can easily access and modify the data, even without using R. These functions below with a prefix \code{sys_grab_xls} provides the first steps to analysis the organized data structure in the \code{.xlsx} file.

The functions are as follow:
\itemize{
\item sys_grab_xlsfactor: Obtain sheets that will form \code{factor} matrix
\item sys_grab_xlsitem: Obtain sheets that will form \code{item} matrix
\item sys_grab_xlsdata: Obtain sheets that will form \code{data} matrix
}
}
\keyword{internal}
