% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_msg_tools.R
\name{sys_message_columns}
\alias{sys_message_columns}
\alias{sys_message_class}
\alias{sys_message_set}
\alias{sys_message_sheet}
\alias{sys_message_code}
\title{System tools: Creating appropriate abort messages}
\usage{
sys_message_columns(columns)

sys_message_class(class)

sys_message_set(set)

sys_message_sheet(sheet)

sys_message_code(code)
}
\arguments{
\item{columns}{Column names as a vector of characters}

\item{class}{Classes as a vector of characters}

\item{set}{Dataset names as a vector of characters}

\item{sheet}{xlsx sheet names as a vector of characters}

\item{code}{Code as a vector of characters}
}
\value{
A single, formatted character string, that can be used directly
}
\description{
An abort or warning message may involve names of many different objects, all need to be processed and collapsed appropriately for people to understand clearly. These functions below help to collapse different object names into 1 string.
}
\examples{
snd:::sys_message_columns(c("a", "b"))
snd:::sys_message_class(c("data.frame", "list"))
snd:::sys_message_set(c("eco", "water"))
snd:::sys_message_sheet(c("#data_eco", "#data_water"))
}
\keyword{internal}
