% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFile_xlsx.R
\name{checkFile_xlsx}
\alias{checkFile_xlsx}
\title{System tools: Check if a file exists}
\usage{
checkFile_xlsx(xlsxFile)
}
\arguments{
\item{xlsxFile}{Directory of the xlsxFile}
}
\value{
A standard message
}
\description{
It checks if a file directory is \code{character(0)}, not existing, or has incorrect extensions
}
\examples{
checkFile_xlsx(xlsxFile)
}
\keyword{internal}
