% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRW_matrix.R
\name{checkRW_matrix}
\alias{checkRW_matrix}
\alias{checkRW_matrix.snd_data}
\alias{checkRW_matrix.snd_item}
\alias{checkRW_matrix.snd_factor}
\title{System tools: Precheck for r/w data matrix}
\usage{
checkRW_matrix(mtx, mtxName)

\method{checkRW_matrix}{snd_data}(mtx, mtxName)

\method{checkRW_matrix}{snd_item}(mtx, mtxName)

\method{checkRW_matrix}{snd_factor}(mtx, mtxName)
}
\arguments{
\item{mtx}{The matrix}

\item{mtxName}{The name of the matrix}
}
\value{
A message
}
\description{
Prior to read-in a matrix, the matrix must loosely resemble the final snd product. These functions check if the matrix fits those criteria to be transformed into a snd matrix.
}
\keyword{Internal}
