% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_key1mtx.R
\name{format_key}
\alias{format_key}
\alias{format_key.sndkey_type}
\alias{format_key.sndkey_item}
\alias{format_key.sndkey_datatype}
\alias{format_key.sndkey_datatype.snd_factor}
\alias{format_key.sndkey_datatype.snd_item}
\alias{format_key.sndkey_factor}
\alias{format_key.sndkey_label}
\alias{format_key.sndkey_label.snd_factor}
\alias{format_key.sndkey_label.snd_item}
\title{System tools: Checking and Formatting the keys during read-in.}
\usage{
format_key(key, mtx, mtxName)

\method{format_key}{sndkey_type}(key, mtx, mtxName)

\method{format_key}{sndkey_item}(key, mtx, mtxName)

format_key.sndkey_datatype(key, mtx, mtxName)

\method{format_key}{sndkey_datatype.snd_factor}(key, mtx, mtxName)

\method{format_key}{sndkey_datatype.snd_item}(key, mtx, mtxName)

\method{format_key}{sndkey_factor}(key, mtx, mtxName)

format_key.sndkey_label(key, mtx, mtxName)

\method{format_key}{sndkey_label.snd_factor}(key, mtx, mtxName)

\method{format_key}{sndkey_label.snd_item}(key, mtx, mtxName)
}
\arguments{
\item{key}{The key in \code{character}. Only 1 key will be accepted.}

\item{mtx}{The matrix in \code{data.frame}. The matrix must contain the key.}

\item{mtxName}{The name of the matrix in character.}
}
\value{
A matrix, checked (or/and modified) by the key.
}
\description{
In SND, a \code{key} is a column within the matrix that can be used to check and modify itself, it's matrix and other matrixes. This function checks and modifies it's own matrix only, using a key. To modify other matrixes, check the function \code{snd:::format_key2mtx}. To enhance security, only 1 key will be accepted.
}
\keyword{internal}
