% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_matrix_by.R
\name{formatData_byItem}
\alias{formatData_byItem}
\alias{formatData_byFactor}
\title{Title}
\usage{
formatData_byItem(mtxData, mtxItem, mtxName)

formatData_byFactor(mtxData, mtxFactor, mtxName)
}
\arguments{
\item{mtxName}{}
}
\value{
A formatted Data matrix
}
\description{
Title
}
\keyword{internal}
