% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatRI_key1mtx.R
\name{formatRI_key}
\alias{formatRI_key}
\title{System tools: Checking and Formatting the keys during read-in.}
\usage{
formatRI_key(key, mtx, mtxName)
}
\arguments{
\item{key}{The key in \code{character}. Only 1 key will be accepted.}

\item{mtx}{The matrix in \code{data.frame}. The matrix must contain the key.}

\item{mtxName}{The name of the matrix in character.}
}
\value{
A matrix, checked (or/and modified) by the key.
}
\description{
In SND, a \code{key} is a column within the matrix that can be used to check and modify itself, it's matrix and other matrixes. This function checks and modifies it's own matrix only, using a key. To modify other matrixes, check the function \code{snd:::formatRI_key2mtx}. To enhance security, only 1 key will be accepted.
}
\keyword{internal}
