% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatRI_matrix.R, R/formatRI_tools.R
\name{formatRI_matrix}
\alias{formatRI_matrix}
\alias{formatRI_matrix.snd_data}
\alias{formatRI_matrix.snd_item}
\alias{formatRI_matrix.snd_factor}
\alias{formatRI_MakeKeyType}
\alias{formatRI_shuffle}
\title{System tools: Formatting while doing read-in operations}
\usage{
formatRI_matrix(mtx, mtxName)

\method{formatRI_matrix}{snd_data}(mtx, mtxName)

\method{formatRI_matrix}{snd_item}(mtx, mtxName)

\method{formatRI_matrix}{snd_factor}(mtx, mtxName)

formatRI_MakeKeyType(mtx, mtxName)

formatRI_shuffle(mtx, mtxName)
}
\arguments{
\item{mtx}{The matrix}

\item{mtxName}{The name of the matrix}
}
\value{
A dataframe
}
\description{
These functions process the matrix and form them into the final product prior to formating them into a dataset.
}
\keyword{internal}
