% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatMtx_shuffle.R, R/formatRI_MakeKeyType.R,
%   R/formatRI_matrix.R
\name{format_shuffle}
\alias{format_shuffle}
\alias{formatRI_MakeKeyType}
\alias{formatRI_matrix}
\title{System tools: Formatting while doing read-in operations}
\usage{
format_shuffle(mtx, mtxName)

formatRI_MakeKeyType(mtx, mtxName)

formatRI_matrix(mtx, mtxName)
}
\arguments{
\item{mtx}{The matrix}

\item{mtxName}{The name of the matrix}
}
\value{
A dataframe
}
\description{
These functions process the matrix and form them into the final product prior to formating them into a dataset.
}
\keyword{internal}
