% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatWO_key1mtx.R
\name{formatWO_key}
\alias{formatWO_key}
\title{Title}
\usage{
formatWO_key(key, mtx, mtxName)
}
\arguments{
\item{mtxName}{}
}
\value{
A matrix, checked (or/and modified) by the key.
}
\description{
Title
}
\keyword{internal}
