% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_snd}
\alias{is_snd}
\alias{is_snd_set}
\alias{is_snd_factor}
\alias{is_snd_item}
\alias{is_snd_data}
\title{Test if an object is an object from \code{snd}}
\usage{
is_snd(x)

is_snd_set(x)

is_snd_factor(x)

is_snd_item(x)

is_snd_data(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value. \code{TRUE} for SED data.
}
\description{
Returns \code{TRUE} for SED data, and \code{FALSE} for all other objects, including lists.
}
\examples{
is_snd(x)
}
