% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_priority.R
\name{key_priority}
\alias{key_priority}
\title{Obtain key execution priority}
\usage{
key_priority(key)
}
\arguments{
\item{key}{A string, with the class prefixed as "sndkey_"}
}
\value{
A number. The higher the number, the higher the priority, the earlier it will execute.
}
\description{
Obtain key execution priority
}
\keyword{internal}
