% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-abort_mtx.R,
%   R/system-abort_mtxColContainNonFactor.R,
%   R/system-abort_mtxColGrpItemsNotSame.R,
%   R/system-abort_mtxColGrpItemsNotSameLength.R,
%   R/system-abort_mtxColGrpItemsNotUnique.R,
%   R/system-abort_mtxColNotAllFilled.R,
%   R/system-abort_mtxColNotAllFilledEmptied.R,
%   R/system-abort_mtxColUnsupport.R, R/system-abort_mtxColWrongClass.R,
%   R/system-abort_mtxKeyDatatypeUnsupported.R,
%   R/system-abort_mtxMissingFactor.R, R/system-abort_mtxMissingItem.R,
%   R/system-abort_mtxMissingKey.R, R/system-abort_mtxWrongClass.R
\name{sys_abort_mtxDuplicatedColumn}
\alias{sys_abort_mtxDuplicatedColumn}
\alias{sys_abort_mtxColContainNonFactor}
\alias{sys_abort_mtxColGrpItemsNotSame}
\alias{sys_abort_mtxColGrpItemsNotSameLength}
\alias{sys_abort_mtxColGrpItemsNotUnique}
\alias{sys_abort_mtxColNotAllFilled}
\alias{sys_abort_mtxColNotAllFilledEmptied}
\alias{sys_abort_mtxColUnsupport}
\alias{sys_abort_mtxColWrongClass}
\alias{sys_abort_mtxKeyDatatypeUnsupported}
\alias{sys_abort_mtxMissingFactor}
\alias{sys_abort_mtxMissingItem}
\alias{sys_abort_mtxMissingKey}
\alias{sys_abort_mtxWrongClass}
\title{System tools: Standard abort message for matrices}
\usage{
sys_abort_mtxDuplicatedColumn(x, name)

sys_abort_mtxColContainNonFactor(x, name, columns)

sys_abort_mtxColGrpItemsNotSame(x, name, columns, group_by, failed_groups)

sys_abort_mtxColGrpItemsNotSameLength(
  x,
  name,
  columns,
  group_by,
  failed_groups
)

sys_abort_mtxColGrpItemsNotUnique(x, name, columns, group_by, failed_groups)

sys_abort_mtxColNotAllFilled(x, name, columns)

sys_abort_mtxColNotAllFilledEmptied(x, name, columns)

sys_abort_mtxColUnsupport(x, name, columns)

sys_abort_mtxColWrongClass(x, name, columns, expected)

sys_abort_mtxKeyDatatypeUnsupported(x, name, unsupport_datatype)

sys_abort_mtxMissingFactor(x, name)

sys_abort_mtxMissingItem(x, name)

sys_abort_mtxMissingKey(x, keys_missing, name)

sys_abort_mtxWrongClass(x, name)
}
\arguments{
\item{x}{The argument holding the matrix}

\item{name}{Optional: name of the matrix}
}
\value{
A standard abort message
}
\description{
\itemize{
\item sys_abort_mtxDuplicatedColumn: There are duplicated column names in the dataframe
\item sys_abort_mtxWrongClass: The input item is not a dataframe
\item sys_abort_mtxMissingKey: There are missing Keys in the dataframe
\item sys_abort_mtxColWrongClass: There are columns with wrong class in the matrix
}
}
\keyword{internal}
