% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_abort_mtx.R,
%   R/sys_abort_mtxColContainFactor.R, R/sys_abort_mtxColContainNonFactor.R,
%   R/sys_abort_mtxColDuplicated.R, R/sys_abort_mtxColGrpItemsNotExclusive.R,
%   R/sys_abort_mtxColGrpItemsNotSame.R,
%   R/sys_abort_mtxColGrpItemsNotSameLength.R,
%   R/sys_abort_mtxColGrpItemsNotUnique.R, R/sys_abort_mtxColNotAllFilled.R,
%   R/sys_abort_mtxColNotAllFilledEmptied.R, R/sys_abort_mtxColUnsupport.R,
%   R/sys_abort_mtxColWrongClass.R, R/sys_abort_mtxKeyDatatypeUnsupported.R,
%   R/sys_abort_mtxKeyLabelIncorrectlyDescribeData.R,
%   R/sys_abort_mtxMissingFactor.R, R/sys_abort_mtxMissingItem.R,
%   R/sys_abort_mtxMissingKey.R, R/sys_abort_mtxMissingSelectedFactor.R,
%   R/sys_abort_mtxMissingSelectedItem.R, R/sys_abort_mtxWrongClass.R
\name{sys_abort_mtx}
\alias{sys_abort_mtx}
\alias{sys_abort_mtxColContainFactor}
\alias{sys_abort_mtxColContainNonFactor}
\alias{sys_abort_mtxColDuplicated}
\alias{sys_abort_mtxColGrpItemsNotExclusive}
\alias{sys_abort_mtxColGrpItemsNotSame}
\alias{sys_abort_mtxColGrpItemsNotSameLength}
\alias{sys_abort_mtxColGrpItemsNotUnique}
\alias{sys_abort_mtxColNotAllFilled}
\alias{sys_abort_mtxColNotAllFilledEmptied}
\alias{sys_abort_mtxColUnsupport}
\alias{sys_abort_mtxColWrongClass}
\alias{sys_abort_mtxKeyformatUnsupported}
\alias{sys_abort_mtxKeyLabelIncorrectlyDescribeData}
\alias{sys_abort_mtxMissingFactor}
\alias{sys_abort_mtxMissingItem}
\alias{sys_abort_mtxMissingSelectedKey}
\alias{sys_abort_mtxMissingSelectedFactor}
\alias{sys_abort_mtxMissingSelectedItem}
\alias{sys_abort_mtxWrongClass}
\title{System tools: Standard abort message for matrices}
\usage{
sys_abort_mtx(name)

sys_abort_mtxColContainFactor(x, name, columns)

sys_abort_mtxColContainNonFactor(x, name, columns)

sys_abort_mtxColDuplicated(x, name)

sys_abort_mtxColGrpItemsNotExclusive(
  x,
  name,
  columns,
  group_by,
  failed_groups,
  exclusive_item
)

sys_abort_mtxColGrpItemsNotSame(x, name, columns, group_by, failed_groups)

sys_abort_mtxColGrpItemsNotSameLength(
  x,
  name,
  columns,
  group_by,
  failed_groups
)

sys_abort_mtxColGrpItemsNotUnique(x, name, columns, group_by, failed_groups)

sys_abort_mtxColNotAllFilled(x, name, columns)

sys_abort_mtxColNotAllFilledEmptied(x, name, columns)

sys_abort_mtxColUnsupport(x, name, columns)

sys_abort_mtxColWrongClass(x, name, columns, expected)

sys_abort_mtxKeyformatUnsupported(x, name, unsupport_format)

sys_abort_mtxKeyLabelIncorrectlyDescribeData(x, name, failed_columns)

sys_abort_mtxMissingFactor(x, name)

sys_abort_mtxMissingItem(x, name)

sys_abort_mtxMissingSelectedKey(x, keys_missing, name)

sys_abort_mtxMissingSelectedFactor(x, factor_missing, name)

sys_abort_mtxMissingSelectedItem(x, item_missing, name)

sys_abort_mtxWrongClass(x, name)
}
\arguments{
\item{name}{Optional: name of the matrix}

\item{x}{The argument holding the matrix}
}
\value{
A standard abort message
}
\description{
System tools: Standard abort message for matrices
}
\keyword{internal}
