% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_support.R
\name{sys_format_support}
\alias{sys_format_support}
\title{List out all supported formats in SND}
\usage{
sys_format_support(with_abbr = F)
}
\value{
A list of supported formats/ classes, in \code{character}
}
\description{
List out all supported formats in SND
}
\examples{
sys_format_support()
}
\keyword{internal}
