% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sndkey_format_support.R
\name{sys_format_support}
\alias{sys_format_support}
\title{List out all supported formats in SND}
\usage{
sys_format_support(with_abbr = FALSE)
}
\arguments{
\item{with_abbr}{A logical value. If \code{TRUE}, returns with the alias, abbreviations, and default values for each type of data. By default (\code{FALSE}), returns none of the above.}
}
\value{
A list of supported formats/ classes, in \code{character}
}
\description{
List out all supported formats in SND
}
\examples{
snd:::sys_format_support()
}
\keyword{internal}
